requirements.txt
python-telegram-bot==21.*
supabase
yt-dlp
ffmpeg-python
opencv-python-headless
instaloader
numpy
gallery-dl